Do more implicit sign change management

+++ src/dps8/dps8_mt.c
@@ -94,7 +94,7 @@ static t_stat mt_rewind (UNIT * uptr, int32 value, char * cptr, void * desc);
 static t_stat mt_show_nunits (FILE *st, UNIT *uptr, int val, void *desc);
 static t_stat mt_set_nunits (UNIT * uptr, int32 value, char * cptr, void * desc);
 static int mt_iom_cmd (UNIT * unitp, pcw_t * p, word12 * stati, bool * need_data, bool * is_read);
-static int mt_iom_io (UNIT * unitp, int chan, int dev_code, uint * tally, uint * cp, word36 * wordp, word12 * stati);
+static int mt_iom_io (UNIT * unitp, uint chan, uint dev_code, uint * tally, uint * cp, word36 * wordp, word12 * stati);
 
 #define N_MT_UNITS_MAX 16
 #define N_MT_UNITS 1 // default
@@ -556,7 +556,7 @@ static int extractWord36FromBuffer (uint8 * bufp, t_mtrlnt tbc, uint * words_pro
     return 0;
   }
 
-static int mt_iom_io (UNIT * unitp, int chan, int dev_code, uint * tally, uint * cp, word36 * wordp, word12 * stati)
+static int mt_iom_io (UNIT * unitp, uint chan, uint dev_code, uint * tally, uint * cp, word36 * wordp, word12 * stati)
   {
     //sim_debug (DBG_DEBUG, & tape_dev, "%s\n", __func__);
     int mt_unit_num = MT_UNIT_NUM (unitp);
@@ -687,7 +687,7 @@ static t_stat mt_set_nunits (UNIT * uptr, int32 value, char * cptr, void * desc)
     int n = atoi (cptr);
     if (n < 1 || n > N_MT_UNITS_MAX)
       return SCPE_ARG;
-    tape_dev . numunits = n;
+    tape_dev . numunits = (uint32) n;
     return SCPE_OK;
   }
